Added logic for Model retraining if the metrics are less than 70%
