+++ Dockerfile
@@ -5,10 +5,7 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     build-essential \
     && apt-get clean \
     && rm -rf /var/lib/apt/lists/*
-
-# Create the /app/final_models directory and set permissions
-RUN mkdir -p /app/final_models && chmod -R 777 /app/final_models
-
+    
 # Switch back to the airflow user
 USER airflow
 COPY requirements.txt /requirements.txt
