--- dags/src/final_model/results_20241108-175047.json
@@ -1,16 +1,0 @@
-{
-    "accuracy": 0.8037155811124627,
-    "precision": 0.8997007229923896,
-    "recall": 0.8037155811124627,
-    "f1_score": 0.832829148781435,
-    "best_params": {
-        "n_estimators": 100,
-        "max_depth": 10,
-        "min_samples_split": 4,
-        "min_samples_leaf": 2,
-        "max_features": "log2",
-        "bootstrap": true
-    },
-    "run_id": "c0310e2e72bf49f685fe7b4939bf4199",
-    "timestamp": "20241108-175047"
-}
