+++ dags/src/final_model/results_20241111-201458.json
@@ -1,0 +1,16 @@
+{
+    "accuracy": 0.8389914851266173,
+    "precision": 0.8988997135008588,
+    "recall": 0.8389914851266173,
+    "f1_score": 0.8586321428556778,
+    "best_params": {
+        "bootstrap": 0,
+        "max_depth": 15.0,
+        "max_features": 0,
+        "min_samples_leaf": 1.0,
+        "min_samples_split": 4.0,
+        "n_estimators": 50.0
+    },
+    "run_id": "10c8fffef983472c905626402d003c96",
+    "timestamp": "20241111-201458"
+}
