+++ dags/src/model_development_pipeline.py
@@ -2,9 +2,10 @@ from airflow import DAG
 from airflow.operators.python_operator import PythonOperator
 from datetime import datetime, timedelta
 import os
+import logging
 
 # Import the function from your script
-from src.Model_Pipeline.model_development_with_mlflow import run_model_development
+from src.Model_Pipeline.model_development_and_evaluation_with_mlflow import run_model_development
 
 default_args = {
     'owner': 'MLopsProjectGroup6',
@@ -22,14 +23,29 @@ dag_2 = DAG(
     catchup=False,
 )
 
-def run_model_development_task():
+def run_model_development_task(**kwargs):
     TRAIN_PATH = "/opt/airflow/data/processed/smote_resampled_train_data.csv"
     TEST_PATH = "/opt/airflow/data/processed/test_data.csv"
-    run_model_development(TRAIN_PATH, TEST_PATH)
+    
+    # Run the model development process
+    final_metrics = run_model_development(TRAIN_PATH, TEST_PATH, max_attempts=3)
+    
+    # Log the final metrics
+    logging.info(f"Final model metrics: {final_metrics}")
+    
+    # You can also push the metrics to XCom if you want to use them in downstream tasks
+    kwargs['ti'].xcom_push(key='final_metrics', value=final_metrics)
+    
+    # Check if the model met the performance threshold
+    if all(metric >= 0.7 for metric in final_metrics.values()):
+        logging.info("Model development successful: All metrics are above 0.7")
+    else:
+        logging.warning("Model development completed, but not all metrics are above 0.7")
 
 model_development_task = PythonOperator(
     task_id='run_model_development',
     python_callable=run_model_development_task,
+    provide_context=True,  # This allows the function to receive kwargs
     dag=dag_2,
 )
